### 2019-06-30 #1538 @whedon
[PRE REVIEW]: pyMannKendall: a python package for non parametric Mann Kendall family of trend tests.
**Submitting author:** @mmhs013 (<a href="http://orcid.org/0000-0002-5361-0633">Md. Manjurul Hussain</a>)
**Repository:** <a href="https://github.com/mmhs013/pyMannKendall" target ="_blank">https://github.com/mmhs013/pyMannKendall</a>
**Version:** v 1.0
