### 2019-07-02 #1538 @mbobra
👋 @dhhagan @molgor Would you be willing to review this submission for JOSS?
