### 2019-07-03 #1538 @molgor
Hi Mónica,

I cannot do this at the moment.  I'll submit my PhD thesis in two months.

Sorry about that.

Best wishes,

Juan

On Tue, Jul 2, 2019 at 6:23 PM Monica Bobra <notifications@github.com>
wrote:

> 👋 @dhhagan <https://github.com/dhhagan> @molgor
> <https://github.com/molgor> Would you be willing to review this
> submission for JOSS?
>
