### 2019-07-03 #1538 @mbobra
@dhhagan Awesome, thanks! Yes, next week is totally fine (and you can take longer than that if needed). Once I find another reviewer, I'll assign you both at the same time.
