### 2019-07-04 #1538 @mmhs013
@mbobra, Mann Kendall is a statistical trend test which is mostly used in climate science, atmospheric science, hydrology etc. So, I think @Chilipp, @andreas-h, @sgrieve, @dvalters might be familiar with this functions. You can invite them for review this submission.
