### 2019-07-04 #1538 @dvalters
Sorry too busy with other things at the moment, not familiar with that
particular test.

On Thu, 4 Jul 2019, 05:07 Md. Manjurul Hussain Shourov, <
notifications@github.com> wrote:

> @mbobra <https://github.com/mbobra>, Mann Kendall is a statistical trend
> test which is mostly used in climate science, atmospheric science,
> hydrology etc. So, I think @Chilipp <https://github.com/Chilipp>,
> @andreas-h <https://github.com/andreas-h>, @sgrieve
> <https://github.com/sgrieve>, @dvalters <https://github.com/dvalters>
> might be familiar with this functions. You can invite them for review this
> submission.
>
