### 2019-07-04 #1538 @andreas-h
Sorry, I am not familiar enough with this in order to ensure a proper review.
