### 2019-07-10 mmhs013/pyMannKendall#1 @rreinecke
[JOSS review] Install instructions
`python setup.py install` will not works without root access of most systems. This could confuse inexperienced users. Maybe add

-  sudo or #
- or add prefix or local install option
