### 2019-07-10 mmhs013/pyMannKendall#2 @rreinecke
[JOSS review] A statement of need
Please state for what the software can be used more clearly in the project description. Maybe even include a plot of an example time-series.
