### 2019-07-10 mmhs013/pyMannKendall#1 @rreinecke
Please also add the dependencies of your package to the project page
