### 2019-07-10 mmhs013/pyMannKendall#3 @rreinecke
[Joss review] Example usage
Please provide a real world example that shows the usage of your tool. Great would be a plot. Also relates to #2
