### 2019-07-10 mmhs013/pyMannKendall#4 @rreinecke
[JOSS review] Functionality documentation
Please compile a documentation using Sphinx or a similar tool. Alternatively, extend the project description with a more detailed description of the method signatures. You already have nice docstrings. A more detailed overview without looking into the code would be nice.
