### 2019-07-10 mmhs013/pyMannKendall#5 @rreinecke
[JOSS review] Automated tests
Please describe how a user can run the tests. Also add a more detailed description what the tests do.
Having only a quick look at your tests it seems like you just comparing values based on some example series. Please consider writing more robust unit tests. If you need more guidance on this feel free to ask.
