### 2019-07-10 mmhs013/pyMannKendall#6 @rreinecke
[JOSS review] References
Please consider adding a DOI in your paper bib to all reference where one is available.
