### 2019-07-10 #1556 @rreinecke
@mmhs013 I did not have the time to take a closer look at your code but here are already some comments on the overall documentation that you could improve:

-  mmhs013/pyMannKendall#1
-  mmhs013/pyMannKendall#2
-  mmhs013/pyMannKendall#3
-  mmhs013/pyMannKendall#4
-  mmhs013/pyMannKendall#5
-  mmhs013/pyMannKendall#6
