### 2019-07-11 #1556 @rreinecke
@mmhs013 I had a look at your code now and I think it is nicely documented but without robust unit tests and an thoroughly explained example problem it is not possible for me to judge if the code actually provides the promised functionality. See also mmhs013/pyMannKendall#5 Let me know if you need further input. I will continue my review when this is implemented.
