### 2019-07-11 17401da
Add dependencies list and update installation instructions according to issue #1
