### 2019-07-11 cfa63a5
Add available doi in bibliography file according to issue #6
