+++ Paper/paper.bib
@@ -12,6 +12,7 @@
   pages={245--259},
   year={1945},
   publisher={JSTOR}
+  doi={10.2307/1907187}
 }
 @article{kendall1975rank,
   title={Rank correlation measures},
@@ -29,6 +30,7 @@
   pages={148--158},
   year={2016},
   publisher={Elsevier}
+  doi={10.1016/j.atmosres.2016.02.008}
 }
 @article{hirsch1982techniques,
   title={Techniques of trend analysis for monthly water quality data},
@@ -39,6 +41,7 @@
   pages={107--121},
   year={1982},
   publisher={Wiley Online Library}
+  doi={10.1029/WR018i001p00107}
 }
 @article{hamed1998modified,
   title={A modified Mann-Kendall trend test for autocorrelated data},
@@ -49,6 +52,7 @@
   pages={182--196},
   year={1998},
   publisher={Elsevier}
+  doi={10.1016/S0022-1694(97)00125-X}
 }
 @article{cox1955some,
   title={Some quick sign tests for trend in location and dispersion},
@@ -59,6 +63,7 @@
   pages={80--95},
   year={1955},
   publisher={JSTOR}
+  doi={10.2307/2333424}
 }
 @article{yue2004mann,
   title={The Mann-Kendall test modified by effective sample size to detect trend in serially correlated hydrological series},
@@ -69,6 +74,7 @@
   pages={201--218},
   year={2004},
   publisher={Springer}
+  doi={10.1023/B:WARM.0000043140.61082.60}
 }
 @article{yue2002applicability,
   title={Applicability of prewhitening to eliminate the influence of serial correlation on the Mann-Kendall test},
@@ -79,6 +85,7 @@
   pages={4--1},
   year={2002},
   publisher={Wiley Online Library}
+  doi={10.1029/2001WR000861}
 }
 @article{yue2002influence,
   title={The influence of autocorrelation on the ability to detect trend in hydrological series},
@@ -89,16 +96,7 @@
   pages={1807--1829},
   year={2002},
   publisher={Wiley Online Library}
-}
-@article{hirsch1982techniques,
-  title={Techniques of trend analysis for monthly water quality data},
-  author={Hirsch, Robert M and Slack, James R and Smith, Richard A},
-  journal={Water resources research},
-  volume={18},
-  number={1},
-  pages={107--121},
-  year={1982},
-  publisher={Wiley Online Library}
+  doi={10.1002/hyp.1095}
 }
 @article{helsel2006regional,
   title={Regional Kendall test for trend},
@@ -109,6 +107,7 @@
   pages={4066--4073},
   year={2006},
   publisher={ACS Publications}
+  doi={10.1021/es051650b}
 }
 @book{hipel1994time,
   title={Time series modelling of water resources and environmental systems},
@@ -126,6 +125,7 @@
   pages={71--84},
   year={2002},
   publisher={Wiley Online Library}
+  doi={10.1002/env.507}
 }
 @inproceedings{theil1950rank,
   title={A rank-invariant method of linear and polynominal regression analysis (Parts 1-3)},
@@ -144,4 +144,5 @@
   pages={1379--1389},
   year={1968},
   publisher={Taylor \& Francis Group}
+  doi={10.1080/01621459.1968.10480934}
 }
