### 2019-07-11 6600b24
Update readme file with extending the project description and add Example according to issue #2, #3 and #4
