+++ README.md
@@ -7,11 +7,14 @@
 [![PyPI - Python Version](https://img.shields.io/pypi/pyversions/pymannkendall.svg)](https://pypi.org/project/pymannkendall/)
 [![status](http://joss.theoj.org/papers/14903dbd55343be89105112e585d262a/status.svg)](http://joss.theoj.org/papers/14903dbd55343be89105112e585d262a)
 
-`pyMannkendal` is a pure Python implementation of non-parametric Mann Kendall trend analysis. Currently, this package has 11 Mann-Kendall Tests and 2 sen's slope estimator function. Brief description of functions are below:
+## What is the Mann Kendall Test ?
+The Mann Kendall Trend Test (sometimes called the MK test) is used to analyze time series data for consistently increasing or decreasing trends (monotonic trends). It is a non-parametric test, which means it works for all distributions (i.e. data doesn't have to meet the assumption of normality), but data should have no serial correlation. If data has a serial correlation, it could be affected in significant level (p-value). That's why results might be misinterpreted. To overcome this problem, researchers proposed several modified Mann Kendall tests (Hamed and Rao Modified MK Test, Yue and Wang Modified MK Test, Modified MK test using Pre-Whitening method etc.). Seasonal Mann Kendall test also developed to remove the effect of seasonality.
+
+Mann Kendall Test is a powerful trend test, so several others modified Mann Kendall tests like Multivariate MK Test, Regional MK Test, Correlated MK test, Partial MK Test etc. were developed for the spacial condition. `pyMannkendal` is a pure Python implementation of non-parametric Mann Kendall trend analysis, which bring together almost all types of Mann Kendall Test. Currently, this package has 11 Mann-Kendall Tests and 2 sen's slope estimator function. Brief description of functions are below:
 
 1.	**Original Mann Kendall test (*original_test*):** Original Mann Kendall test is a nonparametric test, which does not consider serial correlation or seasonal effects.
 
-2.	**Hamed and Rao Modified MK Test (*hamed_rao_modification_test*):** This modified MK test proposed by Hamed and Rao (1998) to address serial autocorrelation issues. (Their) They suggested variance correction approach to improve trend analysis. User can consider first n significant lag by insert lag number in this function. By default, it considered all significant lags.
+2.	**Hamed and Rao Modified MK Test (*hamed_rao_modification_test*):** This modified MK test proposed by Hamed and Rao (1998) to address serial autocorrelation issues. They suggested variance correction approach to improve trend analysis. User can consider first n significant lag by insert lag number in this function. By default, it considered all significant lags.
 
 3.	**Yue and Wang Modified MK Test (*yue_wang_modification_test*):** This is also a variance correction method for considered serial autocorrelation proposed by Yue, S., & Wang, C. Y. (2004). User can also set their desired significant n lags for the calculation.
 
@@ -35,7 +38,7 @@
 
 13.	**Seasonal sen's Slope Estimator (*seasonal_sens_slope*):** This method proposed by Hipel (1994) to estimate the magnitude of the monotonic trend, when data has seasonal effects.
 
-# Function details:
+## Function details:
 
 All Mann-kendall test functions has almost similer input parameters. Those are:
 
@@ -57,13 +60,13 @@ And all Mann-kendall tests return a named tuple which contained:
 
 sen's slope function required data vector. seasonal sen's slope also has optional input period, which by default value is 12. Both sen's slope function return only slope value.
 
-# Dependencies
+## Dependencies
 
 For the installation of `pyMannKendall`, the following packages are required:
[…]
