### 2019-07-11 #1556 @mmhs013
@rreinecke *thanks for your comments*. According to your comments, I am already updated readme file with Installation Instructions, list of dependencies (mmhs013/pyMannKendall#1), project description (mmhs013/pyMannKendall#2) and added new details example ipython notebook file with some plots (mmhs013/pyMannKendall#3) and added DOI in paper bib file (mmhs013/pyMannKendall#6).

The input and output parameters of functions are almost the same, which are described in **Function details** section in the readme file. I think, the new updates of readme file and example also describe more clearly about functions (mmhs013/pyMannKendall#4).

I write this unit test using pytest for my own purpose because when I update/modify my code, this test compares those functions with some known results. If I unintentionally make some mistakes, which produce the wrong result, This will warn me. I think this is fine for me (mmhs013/pyMannKendall#5).

I don't use the unit test previously. So I am not an expert in this section. I don't clearly understand how can I improve my unit test. Do you have any suggestions for me?

Please check my updates. If you are satisfied with my updates, then close relevant issues. I am waiting for your respond.
