### 2019-07-11 mmhs013/pyMannKendall#2 @rreinecke
@mmhs013 Really nice! Can you link the example in the main REDME so that interested user can easily find it?
