### 2019-07-11 mmhs013/pyMannKendall#5 @rreinecke
Ok testing is not an easy topic I will try to give a couple of pointers.

So the issue with the test the way you are applying them right now is the following:
*  It is unclear were the numbers that you comparing to came from
*  What if you made a mistake in calculating these numbers?
*  If you used your own software your calculations might be wrong because your software is wrong
*  Without deep knowledge about the dataset it is not possible for somebody else to understand the numbers that you are comparing to
*  What if in the future a collaborator would like to add support for lets say pandas which is a great package. And relies on that your tests are testing the core functionality of your system including being able to detect wrong data types etc.

All in all you should think about replacing the currents tests. Yes I get they can be helpful for you in the beginning but as soon as you compare to a number with more than 2 decimal places you are doing something wrong.

1) Think about if it is possible to create an artificial data series with e.g. numpy that has certain characteristics e.g. no trend at all. 1000 values that are all 0. What should the behavior of your library be? Can you create a series with pseudo random values? e.g. by putting them into a sinus like function etc. etc. there are ways to create interesting data sets that somebody else can directly understand because they are based on two lines of code and some math

2) Think about if your tests should check for edge cases. What if the series has only one entry? Is that valid? What about NaN values? Should there be an error message? If so you can test that!

My suggestions:

1) Add a section to the README: How to tun the unit tests, with description

2) Have a look at how scipy is doing it https://github.com/scipy/scipy/tree/master/scipy/stats/tests
and https://github.com/numpy/numpy/blob/master/doc/TESTS.rst.txt

Or have a look at other packages I'm sure you can improve your tests! Yes it is more work but it will make your software more robust in the end.
