### 2019-07-11 mmhs013/pyMannKendall#5 @rreinecke
Take your time!
