### 2019-07-11 mmhs013/pyMannKendall#2 @mmhs013
@rreinecke I am already link it in Usage section of main readme file.
