### 2019-07-12 9888eb8
rewrite unit test according to #5
