+++ tests/test_pymannkendall.py
@@ -1,21 +1,111 @@
+# In this unit test file, we check all functions with randomly generated No trendy, trendy, arbitrary data. Those results are compire with R package - modifiedmk, fume, rkt, trend.
+
 import os
 import pytest
-import pandas as pd
+import numpy as np
 import pymannkendall as mk
 
 @pytest.fixture
-def rainfalldata():
-    data = pd.read_csv(os.path.join(os.path.dirname(__file__),"data/Monthly Rainfall.csv"))
-    return data
+def NoTrendData():
+    # Generate 360 random value with same number
+    NoTrendData = np.ones(360)*np.random.randint(10)
+    return NoTrendData
+
+@pytest.fixture
+def NoTrend2dData():
+    # Generate 2 dimentional 360 random value with same number
+    NoTrend2dData = np.ones((360,2))*np.random.randint(10)
+    return NoTrend2dData
     
 @pytest.fixture
-def wqdata():
-    data = pd.read_csv(os.path.join(os.path.dirname(__file__),"data/Water Quality.csv"))
-    return data
+def TrendData():
+    # Generate random 360 trendy data with approx. slope 1 
+    TrendData = np.arange(360) + np.random.rand(360)
+    return TrendData
 
-def test_original_test(rainfalldata):
-    result = mk.original_test(rainfalldata.iloc[:,2])
+@pytest.fixture
+def arbitrary_1d_data():
+    # Generate arbitrary 360 data
+    arbitrary_1d_data = np.array([ 32.,  20.,  25., 189., 240., 193., 379., 278., 301.,   0.,   0.,
+        82.,   0.,   4.,  np.nan,  np.nan, 121., 234., 360., 262., 120.,  30.,
+        11.,   1.,   7.,   3.,  31.,  31., 355., 102., 248., 274., 308.,
+        np.nan,   5.,  26.,  11.,  16.,   6.,  48., 388., 539., 431., 272.,
+       404., 186.,   0.,   2.,   0.,   4.,   1.,  54., 272., 459., 235.,
+       164., 365., 135.,   2.,  np.nan,  np.nan,   4.,   0., 128., 210., 163.,
+       446., 225., 462., 467.,  19.,  13.,   0.,   3.,  17., 132., 178.,
+       338., 525., 623., 145.,  31.,  19.,   3.,   0.,  29.,  25.,  87.,
+       259., 756., 486., 180., 292.,  43.,  92.,   1.,   0.,  16.,   2.,
+         0., 130., 253., 594., 111., 273.,  30.,   0.,   4.,   0.,  27.,
+        24.,  41., 292., 378., 499., 265., 320., 227.,   4.,   0.,   4.,
+        14.,   8.,  48., 416., 240., 404., 207., 733., 105.,   0., 112.,
+         0.,  14.,   0.,  30., 140., 202., 289., 159., 424., 106.,   3.,
+         0.,  65.,   3.,  14.,  58., 268., 466., 432., 266., 240.,  95.,
+         1.,   0.,  10.,  26.,   4., 114.,  94., 289., 173., 208., 263.,
+       156.,   5.,   0.,  16.,  16.,  14.,   0., 111., 475., 534., 432.,
+       471., 117.,  70.,   1.,   3.,  28.,   7., 401., 184., 283., 338.,
+       171., 335., 176.,   0.,   0.,  10.,  11.,   9., 140., 102., 208.,
+       298., 245., 220.,  29.,   2.,  27.,  10.,  13.,  26.,  84., 143.,
+       367., 749., 563., 283., 353.,  10.,   0.,   0.,   0.,   0.,   9.,
+       246., 265., 343., 429., 168., 133.,  17.,   0.,  18.,  35.,  76.,
+       158., 272., 250., 190., 289., 466.,  84.,   0.,   0.,   0.,   0.,
[…]
