### 2019-07-12 aa62441
some modification for python 2.7
