### 2019-07-12 a372cf3
add Tests section in readme file according to issue #5
