### 2019-07-12 mmhs013/pyMannKendall#5 @mmhs013
@rreinecke thanks for your suggestion. According to your suggestion, I already rewrite my test code. I create data with -
- the random same value,
- linearly increasing values with random noise and
- arbitrary values

The results are compared with several R packages like modifiedmk, fume, rkt, trend. Scipy also checks their statistical test with R and Matlab (https://github.com/scipy/scipy/blob/master/scipy/stats/tests/test_stats.py). So, I think, this is OK.

I do not clearly understand, how can I add it in my readme file. Because it is only one line to test it.  I find a [package](https://github.com/connorourke/crystal_torture) in JOSS, who added the Tests section in their readme file. So, I simply adopt their approach.
