### 2019-07-20 785df36
fix version problem according to issue #7
