### 2019-07-21 95bdc72
Implement versioneer according to issue #7
