+++ setup.cfg
@@ -1,0 +1,7 @@
+[versioneer]
+VCS = git
+style = pep440
+versionfile_source = pymannkendall/_version.py
+versionfile_build = pymannkendall/_version.py
+tag_prefix = v
+parentdir_prefix = pymannkendall-
