### 2019-07-21 mmhs013/pyMannKendall#2 @rreinecke
@mmhs013 All right I'm closing this issues. I would still recommend to you to copy one example figure to the readme to make it more attractive for possible users and the link a little more prominent.
