### 2019-07-21 mmhs013/pyMannKendall#5 @rreinecke
Great work! I think this is more comprehensible now! Small typo in the description of the testfile ;)
