### 2019-07-21 #1556 @rreinecke
@mbobra I'm done. I think the submission substantially improved throughout the review and is ready to be accepted after @dhhagan concerns have been addressed.

@mmhs013 I would suggest to you to go through the README and code files with a spellchecker there are still a couple typos and grammar mistakes. ;)

Cheers
Robert
