### 2019-07-23 #1556 @dhhagan
@mbobra Just finished up the review - It looks like after the paper is regenerated and the DOI's appear, it is good to go! I agree with @rreinecke that the english could be improved a bit - @mmhs013 Let me know if you would like any help with this and I could probably run through the ReadMe real quick, though I don't want to step on any toes.
