### 2019-07-23 #1556 @mbobra
@rreinecke and @dhhagan Thank you so much for reviewing! We really appreciate your time and effort ☀️

@mmhs013 We're almost there! I have a couple comments:
* Do you have access to someone who can read over the docs and give some feedback on improving the English? If not, let me know and we can help you.
* Can you please [archive your release on Zenodo](https://guides.github.com/activities/citable-code/) to obtain a DOI and then put that in your README.md file?

After these two items I think we're done 🎉
