+++ pymannkendall/pymannkendall.py
@@ -1,11 +1,9 @@
 """
 Created on 05 March 2018
-Update on 30 June 2019
+Update on 23 July 2019
 @author: Md. Manjurul Hussain Shourov
-version: 1.0
+version: 1.1
 Approach: Vectorisation
-
-Remark: Full 11 mk test and 2 sens slope funtion is complete. Need to optimaze and speed up this code,
 """
 
 from __future__ import division
@@ -72,7 +70,7 @@ def __mk_score(x, n):
     return s
 
 	
-# original mann-kendal's variance S calculation
+# original Mann-Kendal's variance S calculation
 def __variance_s(x, n):
     # calculate the unique data
     unique_x = np.unique(x)
@@ -163,7 +161,7 @@ def sens_slope(x):
     """
     This method proposed by Theil (1950) and Sen (1968) to estimate the magnitude of the monotonic trend.
     Input:
-        x:   a one dimentional vector (list, numpy array or pandas series) data
+        x:   a one dimensional vector (list, numpy array or pandas series) data
     Output:
         slope: sen's slope
     Examples
@@ -181,8 +179,8 @@ def seasonal_sens_slope(x, period=12):
     """
     This method proposed by Hipel (1994) to estimate the magnitude of the monotonic trend, when data has seasonal effects.
     Input:
-        x:   a one dimentional vector (list, numpy array or pandas series) data
-		period: seasonal cycle. For monthly data it is 12, weekly data it is 52 (12 is default)
+        x:   a vector (list, numpy array or pandas series) data
+		period: seasonal cycle. For monthly data it is 12, weekly data it is 52 (12 is the default)
     Output:
         slope: sen's slope
     Examples
@@ -210,14 +208,14 @@ def seasonal_sens_slope(x, period=12):
 	
 def original_test(x, alpha = 0.05):
     """
-    This function check Mann-Kendall (MK) test (Mann 1945, Kendall 1975, Gilbert 1987).
+    This function checks the Mann-Kendall (MK) test (Mann 1945, Kendall 1975, Gilbert 1987).
     Input:
-        x: a one dimentional vector (list, numpy array or pandas series) data
+        x: a vector (list, numpy array or pandas series) data
         alpha: significance level (0.05 default)
     Output:
         trend: tells the trend (increasing, decreasing or no trend)
         h: True (if trend is present) or False (if trend is absence)
-        p: p value of the significance test
+        p: p-value of the significance test
         z: normalized test statistics
         Tau: Kendall Tau
         s: Mann-Kendal's score
@@ -245,15 +243,15 @@ def original_test(x, alpha = 0.05):
 
 def hamed_rao_modification_test(x, alpha = 0.05, lag=None):
     """
-    This function check Modified Mann-Kendall (MK) test using Hamed and Rao (1998) method.
+    This function checks the Modified Mann-Kendall (MK) test using Hamed and Rao (1998) method.
     Input:
-        x: a vector of data
+        x: a vector (list, numpy array or pandas series) data
         alpha: significance level (0.05 default)
         lag: No. of First Significant Lags (default None, You can use 3 for considering first 3 lags, which also proposed by Hamed and Rao(1998))
     Output:
         trend: tells the trend (increasing, decreasing or no trend)
         h: True (if trend is present) or False (if trend is absence)
-        p: p value of the significance test
+        p: p-value of the significance test
         z: normalized test statistics
         Tau: Kendall Tau
         s: Mann-Kendal's score
@@ -308,14 +306,14 @@ def hamed_rao_modification_test(x, alpha = 0.05, lag=None):
 
 def yue_wang_modification_test(x, alpha = 0.05, lag=None):
     """
-    Input: This function check Modified Mann-Kendall (MK) test using Yue and Wang (2004) method.
-        x: a vector of data
[…]
