### 2019-07-23 #1556 @mmhs013
@rreinecke, @dhhagan and @mbobra thanks a lot for reviewing my project.

@mbobra, I already checked the README.md file and others with my friend. I think it's been better than before. I also archive my package on Zenodo and add Zenodo DOI in my README.md file.
