### 2019-07-24 #1556 @mbobra
@mmhs013 It looks like there is a small error in paper.bib. Could you add a comma at the end of line 14?
