### 2019-07-24 9cf1804
fix bug in bib file
