+++ Paper/paper.bib
@@ -11,7 +11,7 @@
   journal={Econometrica: Journal of the Econometric Society},
   pages={245--259},
   year={1945},
-  publisher={JSTOR}
+  publisher={JSTOR},
   doi={10.2307/1907187}
 }
 @article{kendall1975rank,
@@ -29,7 +29,7 @@
   volume={176},
   pages={148--158},
   year={2016},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/j.atmosres.2016.02.008}
 }
 @article{hirsch1982techniques,
@@ -40,7 +40,7 @@
   number={1},
   pages={107--121},
   year={1982},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   doi={10.1029/WR018i001p00107}
 }
 @article{hamed1998modified,
@@ -51,7 +51,7 @@
   number={1-4},
   pages={182--196},
   year={1998},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/S0022-1694(97)00125-X}
 }
 @article{cox1955some,
@@ -62,7 +62,7 @@
   number={1/2},
   pages={80--95},
   year={1955},
-  publisher={JSTOR}
+  publisher={JSTOR},
   doi={10.2307/2333424}
 }
 @article{yue2004mann,
@@ -73,7 +73,7 @@
   number={3},
   pages={201--218},
   year={2004},
-  publisher={Springer}
+  publisher={Springer},
   doi={10.1023/B:WARM.0000043140.61082.60}
 }
 @article{yue2002applicability,
@@ -84,7 +84,7 @@
   number={6},
   pages={4--1},
   year={2002},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   doi={10.1029/2001WR000861}
 }
 @article{yue2002influence,
@@ -95,7 +95,7 @@
   number={9},
   pages={1807--1829},
   year={2002},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   doi={10.1002/hyp.1095}
 }
 @article{helsel2006regional,
@@ -106,7 +106,7 @@
   number={13},
   pages={4066--4073},
   year={2006},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
   doi={10.1021/es051650b}
 }
 @book{hipel1994time,
@@ -124,7 +124,7 @@
   number={1},
   pages={71--84},
   year={2002},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   doi={10.1002/env.507}
 }
 @inproceedings{theil1950rank,
@@ -143,6 +143,6 @@
   number={324},
   pages={1379--1389},
   year={1968},
-  publisher={Taylor \& Francis Group}
+  publisher={Taylor \& Francis Group},
   doi={10.1080/01621459.1968.10480934}
 }
