### 2019-07-24 #1556 @mbobra
@openjournals/joss-eics This paper is ready for acceptance! (I could not get Whedon to check the references but it looks fine in the pdf). Nice work @mmhs013 🎉
