### 2019-07-24 #1556 @mmhs013
@mbobra, all references are not visible in pdf!
