### 2019-07-25 #1556 @kyleniemeyer
Hi @mmhs013, one final thing: could you edit the Zenodo archive to make the title match the paper title?
