### 2019-07-25 2719dd7
Update paper.md
Edits to paper
