### 2019-07-25 #1556 @kyleniemeyer
Also, @mmhs013 please merge the two pull requests I just submitted, which fix a few things in the paper and references: https://github.com/mmhs013/pyMannKendall/pull/9 and https://github.com/mmhs013/pyMannKendall/pull/10
