### 2019-07-25 #1556 @mmhs013
@kyleniemeyer, I already update Zenodo title and also merge you two pull requests. Thanks for you help.
