### 2019-07-25 #1556 @kyleniemeyer
Congrats @mmhs013 on your paper's publication in JOSS! Thanks to  @dhhagan and @rreinecke for reviewing, and @mbobra for editing!
