### 2023-01-06 #1556 @SSSCharmCity
It does not appear that the code implements the full partial MannKendall methods described in Libiseller& Grimvall for one or more covariates WITH SEASONALITY.  I.e L&G describe data for d- variables, p-seasons, and n-years.   This code appears to only address data from 2-variables for n-years.   The utility of this helpful code would be significantly increased if the seasonal form of the partial MK test was coded, allowing the user to specify d=2 and p=1 (currently implemented) as a special case.
